Move definition of permanent flags to separate class
