+++ controller-server/src/main/java/com/yahoo/vespa/hosted/controller/persistence/JobControlFlags.java
@@ -4,8 +4,8 @@ package com.yahoo.vespa.hosted.controller.persistence;
 import com.yahoo.concurrent.maintenance.JobControlState;
 import com.yahoo.transaction.Mutex;
 import com.yahoo.vespa.flags.FlagSource;
-import com.yahoo.vespa.flags.Flags;
 import com.yahoo.vespa.flags.ListFlag;
+import com.yahoo.vespa.flags.PermanentFlags;
 
 import java.util.Set;
 
@@ -21,7 +21,7 @@ public class JobControlFlags implements JobControlState {
 
     public JobControlFlags(CuratorDb curator, FlagSource flagSource) {
         this.curator = curator;
-        this.inactiveJobsFlag = Flags.INACTIVE_MAINTENANCE_JOBS.bindTo(flagSource);
+        this.inactiveJobsFlag = PermanentFlags.INACTIVE_MAINTENANCE_JOBS.bindTo(flagSource);
     }
 
     @Override
