+++ flags/src/test/java/com/yahoo/vespa/flags/FlagsTest.java
@@ -4,8 +4,6 @@ package com.yahoo.vespa.flags;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.node.BooleanNode;
-import com.yahoo.vespa.flags.custom.HostResources;
-import com.yahoo.vespa.flags.custom.SharedHost;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
 
@@ -109,17 +107,7 @@ public class FlagsTest {
                 List.of(instance));
     }
 
-    @Test
-    public void testSharedHostFlag() {
-        SharedHost sharedHost = new SharedHost(List.of(new HostResources(
-                4.0, 16.0, 50.0, 0.3,
-                "fast", "local",
-                10)),
-                null);
-        testGeneric(Flags.SHARED_HOST, sharedHost);
-    }
-
-    private <T> void testGeneric(UnboundFlag<T, ?, ?> unboundFlag, T value) {
+    static <T> void testGeneric(UnboundFlag<T, ?, ?> unboundFlag, T value) {
         FlagSource source = mock(FlagSource.class);
         Flag<T, ?> flag = unboundFlag.bindTo(source);
