+++ node-repository/src/test/java/com/yahoo/vespa/hosted/provision/maintenance/DynamicProvisioningMaintainerTest.java
@@ -15,8 +15,8 @@ import com.yahoo.config.provision.NodeType;
 import com.yahoo.config.provision.RegionName;
 import com.yahoo.config.provision.SystemName;
 import com.yahoo.config.provision.Zone;
-import com.yahoo.vespa.flags.Flags;
 import com.yahoo.vespa.flags.InMemoryFlagSource;
+import com.yahoo.vespa.flags.PermanentFlags;
 import com.yahoo.vespa.flags.custom.ClusterCapacity;
 import com.yahoo.vespa.flags.custom.SharedHost;
 import com.yahoo.vespa.hosted.provision.Node;
@@ -105,7 +105,7 @@ public class DynamicProvisioningMaintainerTest {
     @Test
     public void does_not_deprovision_when_preprovisioning_enabled() {
         var tester = new DynamicProvisioningTester().addInitialNodes();
-        tester.flagSource.withListFlag(Flags.PREPROVISION_CAPACITY.id(), List.of(new ClusterCapacity(1, 1, 3, 2, 1.0)), ClusterCapacity.class);
+        tester.flagSource.withListFlag(PermanentFlags.PREPROVISION_CAPACITY.id(), List.of(new ClusterCapacity(1, 1, 3, 2, 1.0)), ClusterCapacity.class);
         Optional<Node> failedHost = tester.nodeRepository.getNode("host2");
         assertTrue(failedHost.isPresent());
 
@@ -117,7 +117,7 @@ public class DynamicProvisioningMaintainerTest {
     @Test
     public void provision_deficit_and_deprovision_excess() {
         var tester = new DynamicProvisioningTester().addInitialNodes();
-        tester.flagSource.withListFlag(Flags.PREPROVISION_CAPACITY.id(),
+        tester.flagSource.withListFlag(PermanentFlags.PREPROVISION_CAPACITY.id(),
                                        List.of(new ClusterCapacity(2, 48, 128, 1000, 10.0),
                                                new ClusterCapacity(1, 16, 24, 100, 1.0)),
                                        ClusterCapacity.class);
@@ -149,7 +149,7 @@ public class DynamicProvisioningMaintainerTest {
         // Makes provisioned hosts 48-128-1000-10
         tester.hostProvisioner.provisionSharedHost("host4");
 
-        tester.flagSource.withListFlag(Flags.PREPROVISION_CAPACITY.id(),
+        tester.flagSource.withListFlag(PermanentFlags.PREPROVISION_CAPACITY.id(),
                 List.of(new ClusterCapacity(2, 1, 30, 20, 3.0)),
                 ClusterCapacity.class);
 
@@ -171,7 +171,7 @@ public class DynamicProvisioningMaintainerTest {
 
         // Add a second cluster equal to the first.  It should fit on existing host3 and hostname100.
 
-        tester.flagSource.withListFlag(Flags.PREPROVISION_CAPACITY.id(),
+        tester.flagSource.withListFlag(PermanentFlags.PREPROVISION_CAPACITY.id(),
                 List.of(new ClusterCapacity(2, 1, 30, 20, 3.0),
                         new ClusterCapacity(2, 1, 30, 20, 3.0)),
                 ClusterCapacity.class);
@@ -185,7 +185,7 @@ public class DynamicProvisioningMaintainerTest {
         // in this test, due to skew), so host3 will be deprovisioned when hostname101 is provisioned.
         // host3 is a 24-64-100-10 while hostname100 is 48-128-1000-10.
 
-        tester.flagSource.withListFlag(Flags.PREPROVISION_CAPACITY.id(),
+        tester.flagSource.withListFlag(PermanentFlags.PREPROVISION_CAPACITY.id(),
                 List.of(new ClusterCapacity(2, 1, 30, 20, 3.0),
                         new ClusterCapacity(2, 24, 64, 100, 1.0)),
                 ClusterCapacity.class);
@@ -201,7 +201,7 @@ public class DynamicProvisioningMaintainerTest {
 
         // If the preprovision capacity is reduced, we should see shared hosts deprovisioned.
 
-        tester.flagSource.withListFlag(Flags.PREPROVISION_CAPACITY.id(),
+        tester.flagSource.withListFlag(PermanentFlags.PREPROVISION_CAPACITY.id(),
                 List.of(new ClusterCapacity(1, 1, 30, 20, 3.0)),
                 ClusterCapacity.class);
 
@@ -264,7 +264,7 @@ public class DynamicProvisioningMaintainerTest {
         var tester = new DynamicProvisioningTester().addInitialNodes();
         tester.hostProvisioner.provisionSharedHost("host4");
 
-        tester.flagSource.withJacksonFlag(Flags.SHARED_HOST.id(), new SharedHost(null, minCount), SharedHost.class);
+        tester.flagSource.withJacksonFlag(PermanentFlags.SHARED_HOST.id(), new SharedHost(null, minCount), SharedHost.class);
         tester.maintainer.maintain();
         assertEquals(provisionCount, tester.hostProvisioner.provisionedHosts.size());
         assertEquals(deprovisionCount, tester.hostProvisioner.deprovisionedHosts);
@@ -289,7 +289,7 @@ public class DynamicProvisioningMaintainerTest {
     public void test_minimum_capacity() {
         var tester = new DynamicProvisioningTester();
         NodeResources resources1 = new NodeResources(24, 64, 100, 10);
-        tester.flagSource.withListFlag(Flags.PREPROVISION_CAPACITY.id(),
+        tester.flagSource.withListFlag(PermanentFlags.PREPROVISION_CAPACITY.id(),
                 List.of(new ClusterCapacity(2, resources1.vcpu(), resources1.memoryGb(), resources1.diskGb(), resources1.bandwidthGbps())),
                 ClusterCapacity.class);
         tester.maintainer.maintain();
@@ -309,7 +309,7 @@ public class DynamicProvisioningMaintainerTest {
         tester.assertNodesUnchanged();
 
         // Must be able to allocate 2 nodes with "no resource requirement"
-        tester.flagSource.withListFlag(Flags.PREPROVISION_CAPACITY.id(),
+        tester.flagSource.withListFlag(PermanentFlags.PREPROVISION_CAPACITY.id(),
                 List.of(new ClusterCapacity(2, 0, 0, 0, 0.0)),
                 ClusterCapacity.class);
 
@@ -330,11 +330,11 @@ public class DynamicProvisioningMaintainerTest {
         tester.assertNodesUnchanged();
 
         // Clearing flag does nothing
-        tester.flagSource.withListFlag(Flags.PREPROVISION_CAPACITY.id(), List.of(), ClusterCapacity.class);
+        tester.flagSource.withListFlag(PermanentFlags.PREPROVISION_CAPACITY.id(), List.of(), ClusterCapacity.class);
         tester.assertNodesUnchanged();
 
         // Increasing the capacity provisions additional hosts
-        tester.flagSource.withListFlag(Flags.PREPROVISION_CAPACITY.id(),
+        tester.flagSource.withListFlag(PermanentFlags.PREPROVISION_CAPACITY.id(),
                 List.of(new ClusterCapacity(3, 0, 0, 0, 0.0)),
                 ClusterCapacity.class);
         assertEquals(0, tester.provisionedHostsMatching(sharedHostNodeResources));
@@ -347,7 +347,7 @@ public class DynamicProvisioningMaintainerTest {
         tester.assertNodesUnchanged();
 
         // Requiring >0 capacity does nothing as long as it fits on the 3 hosts
-        tester.flagSource.withListFlag(Flags.PREPROVISION_CAPACITY.id(),
+        tester.flagSource.withListFlag(PermanentFlags.PREPROVISION_CAPACITY.id(),
                 List.of(new ClusterCapacity(3,
                         resources1.vcpu() - applicationNodeResources.vcpu(),
                         resources1.memoryGb() - applicationNodeResources.memoryGb(),
@@ -357,7 +357,7 @@ public class DynamicProvisioningMaintainerTest {
         tester.assertNodesUnchanged();
 
         // But requiring a bit more in the cluster => provisioning of 2 shared hosts.
-        tester.flagSource.withListFlag(Flags.PREPROVISION_CAPACITY.id(),
+        tester.flagSource.withListFlag(PermanentFlags.PREPROVISION_CAPACITY.id(),
                 List.of(new ClusterCapacity(3,
                         resources1.vcpu() - applicationNodeResources.vcpu() + 1,
                         resources1.memoryGb() - applicationNodeResources.memoryGb() + 1,
